back/requirements.txt
Django==5.0.4
gunicorn==22.0.0
psycopg2==2.9.9
djangorestframework==3.15.1
markdown==3.6
django-filter==24.2
